milleniumfalcon

+++ src/hu/flow/Urhajo.java
@@ -5,6 +5,6 @@ public interface Urhajo {
 
 
     public boolean legyorsuljaE(Urhajo urhajo);
-    public int milyenGyors();
+    public double milyenGyors();
 
 }
